+++ tests/interfaces_ace_test.py
@@ -53,7 +53,7 @@ def test_start(ace_mismatch, tmpdir, monkeypatch):
     with monkeypatch.context() as m:
         m.setattr(ace, 'Popen', popen)
         m.setattr(ace, '_ace_version', lambda x: (0, 9, 29))
-        with pytest.raises(ace.AceProcessError):
+        with pytest.raises(ace.ACEProcessError):
             ace.AceParser(str(grm))
-        with pytest.raises(ace.AceProcessError):
+        with pytest.raises(ace.ACEProcessError):
             ace.parse(str(grm), 'Dogs sleep.')